backend-api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
opencv-python==4.8.0.74
numpy==1.24.3
pillow==9.5.0
torch==2.0.1
ultralytics==8.0.145 